requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
# Updated for Python 3.13 wheel availability
pydantic==2.8.2
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain==0.1.5
langchain-openai==0.0.5
langchain-anthropic==0.1.1
langgraph==0.0.20

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Task Queue
redis==5.0.1
celery==5.3.6

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Type Checking
mypy==1.8.0
types-requests==2.31.0.20240106

# Monitoring
prometheus-fastapi-instrumentator==7.0.0
